--- bls12_381_g1/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "ark-bls12-381-g1"
-version = "0.0.1"
-authors = ["Syed Hosseini <syed@riseup.net>"]
-description = "Reimplementation of BLS12-381 curve without GLV"
-repository.workspace = true
-edition.workspace = true
-
-[dependencies]
-ark-ff = { workspace = true }
-ark-ec = { workspace = true }
-ark-std = { workspace = true }
-ark-bls12-381 = { workspace = true, features = [ "scalar_field", "curve"] }
-
-[dev-dependencies]
-ark-serialize = { workspace = true }
-#ark-algebra-test-templates = { workspace = true }
-#ark-algebra-bench-templates = { workspace = true }
-
-[features]
-default = []
-std = [ "ark-std/std", "ark-ff/std", "ark-ec/std", "ark-bls12-381/std" ]
-asm = [ "ark-ff/asm" ]
-
-# [[bench]]
-# name = "bls12_381_g1"
-# path = "benches/ed_by_bls12_381.rs"
-# harness = false
