--- bls12_381_g1/benches/_trial_temp/test.log
@@ -1,1 +1,0 @@
-2025-11-26 10:12:35-0500 [-] Log opened.
