--- bls12_381_g1/src/curves/built_boost-stamp
@@ -1,1 +1,0 @@
-../../../../../code_ouinet/ouinet/build/boost/src/built_boost-stamp
