--- bls12_381_g1/src/curves/mod.rs
@@ -1,67 +1,0 @@
-use ark_ec::AdditiveGroup;
-use ark_ec::{
-    models::CurveConfig,
-    short_weierstrass::{Affine, Projective, SWCurveConfig},
-};
-use ark_ff::MontFp;
-
-use crate::{Fq, Fr};
-
-#[cfg(test)]
-mod tests;
-
-pub type SWAffine = Affine<SWConfig>;
-pub type SWProjective = Projective<SWConfig>;
-
-/// This is a Prime ordered short Weierstrass curve with prime group order matching of
-/// BLS12-381 curve.
-/// These curves have equations of the form:
-///  y² = x^3 + a*x + b
-/// over some base finite field Fq.
-///
-///
-/// Elliptic Curve defined by y^2 = x^3 +
-/// 1513166797894431528222373717573664845820984622488826599884837603702838646568*x +
-/// 50922708377231758951225366790612300991435319491014185911360037520993381910261
-/// over Finite Field of size
-/// 52435875175126190479447740508185965837256304113503012511244875124696220556829
-///
-/// Sage script to calculate these:
-/// <https://github.com/w3f/research-internal/blob/master/stuff/BLS12-381%20Lollipop.ipynb>
-///
-///
-#[derive(Clone, Default, PartialEq, Eq)]
-pub struct SWConfig;
-
-impl CurveConfig for SWConfig {
-    type BaseField = Fq;
-    type ScalarField = Fr;
-
-    /// COFACTOR = (x - 1)^2 / 3  = 76329603384216526031706109802092473003
-    const COFACTOR: &'static [u64] = &[0x8c00aaab0000aaab, 0x396c8c005555e156];
-
-    /// COFACTOR_INV = COFACTOR^{-1} mod r
-    /// = 52435875175126190458656871551744051925719901746859129887267498875565241663483
-    const COFACTOR_INV: Fr =
-        MontFp!("52435875175126190458656871551744051925719901746859129887267498875565241663483");
-}
-
-impl SWCurveConfig for SWConfig {
-    /// COEFF_A = 1513166797894431528222373717573664845820984622488826599884837603702838646568
-    /// COEFF_A = 0
-    const COEFF_A: Fq = Fq::ZERO;
-
-    /// COEFF_B = 4
-    const COEFF_B: Fq = MontFp!("4");
-
-    /// AFFINE_GENERATOR_COEFFS = (G1_GENERATOR_X, G1_GENERATOR_Y)
-    const GENERATOR: SWAffine = SWAffine::new_unchecked(G1_GENERATOR_X, G1_GENERATOR_Y);
-}
-
-/// G1_GENERATOR_X =
-/// 3685416753713387016781088315183077757961620795782546409894578378688607592378376318836054947676345821548104185464507
-pub const G1_GENERATOR_X: Fq = MontFp!("3685416753713387016781088315183077757961620795782546409894578378688607592378376318836054947676345821548104185464507");
-
-/// G1_GENERATOR_Y =
-/// 1339506544944476473020471379941921221584933875938349620426543736416511423956333506472724655353366534992391756441569
-pub const G1_GENERATOR_Y: Fq = MontFp!("1339506544944476473020471379941921221584933875938349620426543736416511423956333506472724655353366534992391756441569");
