--- bls12_381_g1/src/fields/fr.rs
@@ -1,10 +1,0 @@
-use ark_ff::fields::{Fp256, MontBackend, MontConfig};
-use ark_std::convert::TryInto;
-
-#[derive(MontConfig)]
-#[modulus = "52435875175126190479447740508185965837690552500527637822603658699938581184513"]
-#[generator = "7"]
-#[small_subgroup_base = "3"]
-#[small_subgroup_power = "1"]
-pub struct FrConfig;
-pub type Fr = Fp256<MontBackend<FrConfig, 4>>;
