--- bls12_381_g1/src/fields/mod.rs
@@ -1,14 +1,0 @@
-pub mod fr;
-pub use self::fr::*;
-
-pub mod fq;
-pub use self::fq::*;
-
-pub mod fq2;
-pub use self::fq2::*;
-
-pub mod fq6;
-pub use self::fq6::*;
-
-pub mod fq12;
-pub use self::fq12::*;
