+++ experimental/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "w3f-bls-experimental"
+version = "0.1.0"
+edition = "2018"
+description = "Experimental aggregate BLS signature strategies"
+repository = "https://github.com/w3f/bls"
+
+[dependencies]
+w3f-bls = { path = "..", features = ["std"] }
+ark-ff = { workspace = true }
+ark-ec = { workspace = true }
+ark-serialize = { workspace = true }
+arrayref = { version = "0.3", default-features = false }
+rand = { version = "0.8.5", default-features = false }
+sha3 = { version = "0.10", default-features = false }
+
+[features]
+default = ["std"]
+std = ["rand/std", "w3f-bls/std"]
+
+[[example]]
+name = "aggregated"
+path = "examples/aggregated.rs"
