+++ experimental/sw_by_bls12_381/benches/sw_by_bls12_381.rs
@@ -1,4 +1,4 @@
-//use ark_algebra_bench_templates::*;
+use ark_algebra_bench_templates::*;
 use ark_sw_by_bls12_381::{fq::Fq, fr::Fr, SWProjective as G};
 
 bench!(
