+++ experimental/sw_by_bls12_381/scripts/scalar_field.sage
