--- sw_by_bls12_381/src/curves/tests.rs
@@ -1,5 +1,0 @@
-//use crate::*;
-//use ark_algebra_test_templates::*;
-
-//test_group!(sw; SWProjective; sw);
-//test_group!(te; EdwardsProjective; te);
